- Modificar datos (rematando último retoques) - Agregado elemenetos borrados tras el merge - Usuarios.jsp modificado por error tras el merge

+++ iw/src/main/java/es/fdi/iw/controller/HomeController.java
@@ -948,15 +948,15 @@ public class HomeController {
 		if(id == u.getID() || u.getRol().equals("admin")){
 			cambio = comprobarCampos(edit, nombreUsuario, pass, email, telefono);
 			
-			if((andLogica(1, cambio, 1000000001) == false) && (andLogica(9, cambio, 1000100001) == true)) //1er and es si el valor es erroneo - 2do campo si se ha cambiado el valor
+			if((andLogica(9, cambio, 1000000001) == true) && (andLogica(1, cambio, 1000100001) == false)) //1ercampo si se ha cambiado el valor- 2do and es si el valor es erroneo 
 				edit.setNombre(nombreUsuario);
-			if((andLogica(2, cambio, 1000000010) == false) && (andLogica(10, cambio, 1001000010) == true))
+			if((andLogica(8, cambio, 1000000010) == true) && (andLogica(2, cambio, 1001000010) == false))
 				edit.setHashedAndSalted(edit.generateHashedAndSalted(pass));
-			if((andLogica(3, cambio, 1000000100) == false) && (andLogica(11, cambio, 1010000100) == true))
+			if((andLogica(7, cambio, 1000000100) == true) && (andLogica(3, cambio, 1010000100) == false))
 				edit.setEmail(email);
-			if((andLogica(4, cambio, 1000001000) == false) && (andLogica(12, cambio, 1100001000) == true))
+			if((andLogica(6, cambio, 1000001000) == true) && (andLogica(4, cambio, 1100001000) == false))
 				edit.setTelefono(telefono);
-			/*if(andLogica(8, cambio, 1000010000))
+			/*if(andLogica(5, cambio, 1000010000))
 				edit.setFoto(photo.getOriginalFilename());*/
 		}
 		
@@ -976,11 +976,23 @@ public class HomeController {
 		Usuario u = (Usuario)session.getAttribute("user");
 		
 		
-		if(!photo.isEmpty()){
-			//edit.setFoto(photo.getOriginalFilename());
+		if(!photo.isEmpty() && id == u.getID()){
+			if (!photo.isEmpty()) {
+			    try {
+			        byte[] bytes = photo.getBytes();
+			        BufferedOutputStream stream =
+			                new BufferedOutputStream(
+			                		new FileOutputStream(ContextInitializer.getFile("usuarios", Long.toString(u.getID()) + ".jpg")));
+			        stream.write(bytes);
+			        stream.close();
+					
+			    } catch (Exception e) {
+			    	return new ResponseEntity<String>("Fotografía no adjuntada al usuario satisfactoriamente", HttpStatus.NOT_MODIFIED);	
+			    }
+			 }
 		}
 		
-		return new ResponseEntity<String>("MEC", HttpStatus.OK);
+		return new ResponseEntity<String>("Fotografía adjuntada al usuario satisfactoriamente", HttpStatus.OK);
 		
 			
 	}
@@ -1066,9 +1078,20 @@ public class HomeController {
 	
 	@Transactional
 	@RequestMapping(value = "/eliminarUsuario", method = RequestMethod.POST)
-	public String eliminarUsuario(@RequestParam("idUsuario") long idUsuario,Model model){
+	public String eliminarUsuario(@RequestParam("idUsuario") long idUsuario, Model model, HttpSession session){
+			
+		
 			Usuario usuario= entityManager.find(Usuario.class, idUsuario);
-			entityManager.remove(usuario);			
+			Usuario u = (Usuario)session.getAttribute("user");
+			
+			if(idUsuario == u.getID()){
+				entityManager.remove(usuario);
+			
+				File foto = ContextInitializer.getFile("usuarios", Long.toString(idUsuario));
+				foto.delete();
+				
+			}
+			
 			return "eliminarUsuario";
     }
 
@@ -1336,6 +1359,25 @@ public class HomeController {
 		 }
 		
 		return new ResponseEntity<String>("Fotografía adjuntada al usuario satisfactoriamente", HttpStatus.OK);	
+	}
+	
+	@ResponseBody
+	@RequestMapping(value = "/disponibilidadApodo", method = RequestMethod.POST)
+	public ResponseEntity<String> comprobarDisponibilidadApodo(
+			@RequestParam("apodo") String apodo,
+			HttpSession session){
+		
+		String res = new String("ocupado");
+		
+		try{				
+			String apodoBd = (String) entityManager.createNamedQuery("dameApodoUsuario").setParameter("apodo", apodo).getSingleResult();
+			
+		} catch (NoResultException nre) {
+			res = "libre";
+			return new ResponseEntity<String>(res, HttpStatus.OK);	
+			
+		}
+		return new ResponseEntity<String>(res, HttpStatus.OK);	
 	}
 	
 	@Transactional
